chore(Android): 整理示例页面
