+++ android/hummer-demo-app/src/main/java/com/didi/hummer/demo/HummerSinglePageFragmentActivity.java
@@ -15,13 +15,13 @@ import com.didi.hummer.adapter.navigator.impl.DefaultNavigatorAdapter;
  *
  * Created by XiaoFeng on 2020-01-02.
  */
-public class HummerPageFragmentActivity extends FragmentActivity {
+public class HummerSinglePageFragmentActivity extends FragmentActivity {
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        setContentView(R.layout.activity_page_fragment);
+        setContentView(R.layout.activity_single_page_fragment);
 
         NavPage page = null;
         try {
@@ -29,7 +29,7 @@ public class HummerPageFragmentActivity extends FragmentActivity {
         } catch (Exception e) {
             e.printStackTrace();
         }
-        loadFragment(HummerPageFragment.newInstance(page), R.id.fragment_container, "fragment_hummer");
+        loadFragment(HummerSinglePageFragment.newInstance(page), R.id.fragment_container, "fragment_hummer");
     }
 
     private void loadFragment(Fragment fragment, int fragmentContainerId, String tag) {
