+++ android/hummer-demo-app/src/main/res/layout/activity_hummer_embedded_page.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#F3F4F5"
-    tools:context=".HummerEmbeddedActivity">
+    tools:context=".HummerEmbeddedPageActivity">
 
     <TextView
         android:id="@+id/layout_above"
