+++ android/hummer-demo-app/src/main/res/layout/activity_single_page_fragment.xml
@@ -4,6 +4,6 @@
     android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HummerPageFragmentActivity">
+    tools:context=".HummerSinglePageFragmentActivity">
 
 </FrameLayout>
